requirements.txt
numpy==1.17.2
pandas==0.25.1
torch==1.2.0
torchvision==0.4.0
GitPython==2.1.13
alive-progress==1.3.0
